requirements.txt
Appium-Python-Client~=2.6.0
pytest==7.1.2
